bugfix : 스토리지 이미지 업로드 함수 자체를 클라이언트(유틸함수)로 이동
