+++ src/app/(write)/register/log/page.tsx
@@ -59,7 +59,6 @@ const LogPage = () => {
   };
 
   const onSubmit = async (values: LogFormValues) => {
-    // console.log(values);
     mutate({ values });
   };
