+++ src/app/actions/storage.ts
@@ -2,7 +2,6 @@
 
 import { createClient } from '@/lib/supabase/server';
 import { ApiResponse } from '@/types/api/common';
-import pLimit from 'p-limit';
 import { StorageBucket } from '../../types/api/storage';
 import { getUser } from './user';
 
@@ -64,42 +63,6 @@ export async function getSignedUploadUrl(
   return { ...data, path }; // signedUrl, path
 }
 
-/* signedUrl로 이미지 업로드 */
-type UploadImageOptions = {
-  folder?: string;
-  subfolder?: string;
-  filename: string;
-};
-
-/* 단일 이미지 업로드 */
-export async function uploadImageToSupabase(
-  bucketName: StorageBucket,
-  file: Blob,
-  options: UploadImageOptions
-): Promise<ApiResponse<string>> {
-  try {
-    // 1. signed URL 발급
-    const supabase = await createClient();
-    const { path, token } = await getSignedUploadUrl(
-      bucketName,
-      options.filename,
-      options.folder,
-      options.subfolder
-    );
-
-    // 2. signed URL로 업로드
-    const { data, error } = await supabase.storage
-      .from(bucketName)
-      .uploadToSignedUrl(path, token, file);
-
-    if (error) throw new Error('업로드 실패');
-    return { success: true, data: data?.fullPath };
-  } catch (error) {
-    console.error('Image upload failed:', error);
-    return { success: false, msg: ' 이미지 업로드 실패' };
-  }
-}
-
 /* SignedURLs  */
 export async function getMultipleSignedUploadUrls(
   bucketName: StorageBucket,
@@ -124,74 +87,6 @@ export async function getMultipleSignedUploadUrls(
   } catch (err) {
     console.error('getMultipleSignedUploadUrls 실패:', err);
     return { success: false, msg: 'Signed upload 처리 중 오류 발생' };
-  }
-}
-
-/* 다중 이미지 업로드 */
-type UploadMultipleImagesOptions = {
-  files: Blob[];
-  bucketName: StorageBucket;
-  folder?: string;
-  subfolder?: string;
-};
-
-export async function uploadMultipleImages({
-  files,
-  bucketName,
-  folder,
-  subfolder,
-}: UploadMultipleImagesOptions): Promise<ApiResponse<string[]>> {
-  try {
-    const me = await getUser();
-    if (!me) throw new Error('유저 없음');
-    if (files.length === 0) return { success: true, data: [] };
-
-    const supabase = await createClient();
-    // 1. 업로드할 파일 경로 생성
-    const fileNames = files.map((_, i) => {
-      const filename = `${i}.webp`;
-      return [me.user_id, folder, subfolder, filename].filter(Boolean).join('/');
-    });
-
-    // 2. Signed URL 목록 발급
-    const signedUrlsResult = await getMultipleSignedUploadUrls(bucketName, fileNames);
-    if (!signedUrlsResult.success) {
-      throw new Error(signedUrlsResult.msg);
-    }
-    const signedUrlsData = signedUrlsResult.data;
-
-    const limit = pLimit(3);
-
-    // 3. Signed URL에 이미지 업로드
-    // const uploadPromises = files.map(async (file, i) => {
-    //   const { path, token } = signedUrlsData[i];
-
-    //   const { data, error } = await supabase.storage
-    //     .from(bucketName)
-    //     .uploadToSignedUrl(path, token, file);
-
-    //   if (error) throw new Error(`파일 업로드 실패: ${fileNames[i]}`);
-
-    //   return data?.fullPath;
-    // });
-    const uploadPromises = files.map((file, i) =>
-      limit(async () => {
-        const { path, token } = signedUrlsData[i];
-        const { data, error } = await supabase.storage
-          .from(bucketName)
-          .uploadToSignedUrl(path, token, file);
-
-        if (error) throw new Error(`파일 업로드 실패: ${fileNames[i]}`);
-        return data?.fullPath;
-      })
-    );
-
-    const urls = await Promise.all(uploadPromises);
-
-    return { success: true, data: urls };
-  } catch (error) {
-    console.error('다중 이미지 업로드 실패:', error);
-    return { success: false, msg: '이미지 업로드 중 오류가 발생했습니다.' };
   }
 }
